+++ packages/plugin-packages/mashroom-browser-cache/test/MashroomCacheControlService.test.ts
@@ -24,7 +24,7 @@ describe('MashroomCacheControlService', () => {
         };
 
         const cacheControlService = new MashroomCacheControlService(false, false, 1800, loggerFactory);
-        cacheControlService.addCacheControlHeader('ALWAYS', req, res);
+        cacheControlService.addCacheControlHeader('SHARED', req, res);
 
         expect(cacheControlHeader).toBeTruthy();
         expect(cacheControlHeader).toBe('public, max-age=1800');
@@ -57,7 +57,40 @@ describe('MashroomCacheControlService', () => {
         };
 
         const cacheControlService = new MashroomCacheControlService(false, false, 1800, loggerFactory);
-        cacheControlService.addCacheControlHeader('ALWAYS', req, res);
+        cacheControlService.addCacheControlHeader('SHARED', req, res);
+
+        expect(cacheControlHeader).toBeTruthy();
+        expect(cacheControlHeader).toBe('public, max-age=1800');
+    });
+
+    it('sets the Cache-Control header correctly for an authenticated user for private resources', () => {
+        const req: any = {
+            method: 'GET',
+            pluginContext: {
+                loggerFactory,
+                services: {
+                    security: {
+                        service: {
+                            getUser: () => ({
+                                username: 'test'
+                            })
+                        }
+                    }
+                }
+            }
+        };
+
+        let cacheControlHeader = null;
+        const res: any = {
+            set: (headerName: string, header: string) => {
+                if (headerName === 'Cache-Control') {
+                    cacheControlHeader = header;
+                }
+            }
+        };
+
+        const cacheControlService = new MashroomCacheControlService(false, false, 1800, loggerFactory);
+        cacheControlService.addCacheControlHeader('PRIVATE_IF_AUTHENTICATED', req, res);
 
         expect(cacheControlHeader).toBeTruthy();
         expect(cacheControlHeader).toBe('private, max-age=1800');
@@ -81,7 +114,7 @@ describe('MashroomCacheControlService', () => {
         };
 
         const cacheControlService = new MashroomCacheControlService(false, false, 1800, loggerFactory);
-        cacheControlService.addCacheControlHeader('ALWAYS', req, res);
+        cacheControlService.addCacheControlHeader('SHARED', req, res);
 
         expect(cacheControlHeader).toBeFalsy();
     });
@@ -105,7 +138,7 @@ describe('MashroomCacheControlService', () => {
         };
 
         const cacheControlService = new MashroomCacheControlService(false, true, 1800, loggerFactory);
-        cacheControlService.addCacheControlHeader('ALWAYS', req, res);
+        cacheControlService.addCacheControlHeader('SHARED', req, res);
 
         expect(cacheControlHeader).toBeTruthy();
         expect(cacheControlHeader).toBe('no-cache, no-store, max-age=0');
@@ -129,7 +162,7 @@ describe('MashroomCacheControlService', () => {
         };
 
         const cacheControlService = new MashroomCacheControlService(true, false, 1800, loggerFactory);
-        cacheControlService.addCacheControlHeader('ALWAYS', req, res);
+        cacheControlService.addCacheControlHeader('SHARED', req, res);
 
         expect(cacheControlHeader).toBeTruthy();
         expect(cacheControlHeader).toBe('no-cache, no-store, max-age=0');