+++ packages/plugin-packages/mashroom-browser-cache/type-definitions/index.d.ts
@@ -1,7 +1,7 @@
 
 import type {Request, Response} from 'express';
 
-export type CachingPolicy = 'ALWAYS' | 'NEVER' | 'ONLY_FOR_ANONYMOUS_USERS';
+export type CachingPolicy =  'SHARED' | 'PRIVATE_IF_AUTHENTICATED' | 'NEVER' | 'ONLY_FOR_ANONYMOUS_USERS';
 
 export interface MashroomCacheControlService {
     /**